refactor: 디렉토리 이동 #6
