+++ src/main/java/util/HttpRequestUtils.java
@@ -1,4 +1,4 @@
-package model;
+package util;
 
 import java.util.HashMap;
 import java.util.Map;
