+++ src/main/java/util/IOUtils.java
@@ -1,4 +1,4 @@
-package model;
+package util;
 
 import java.io.BufferedReader;
 import java.io.IOException;
